% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ini.R
\name{create_scenario_ini}
\alias{create_scenario_ini}
\title{Create a new Hector input .ini file for a given scenario. New .ini files are
written to hectordata/inst/input.}
\usage{
create_scenario_ini(scenario)
}
\arguments{
\item{scenario}{Character vector; Name of the scenario for which an input file is being created.}

\item{emission_file}{Character vector; Name of the emission file for the given scenario.}
}
\value{
Character vector; Path of the new .ini file
}
\description{
Create a new Hector input .ini file for a given scenario. New .ini files are
written to hectordata/inst/input.
}
\examples{
create_scenario_ini("rcp45")

}
