% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-utils.R
\name{fetch_minted_data}
\alias{fetch_minted_data}
\title{Download the input data from zenodo}
\usage{
fetch_minted_data(version = "0.0.0.91", write_to = ZENODO_DIR)
}
\arguments{
\item{version}{str value corresponding to the version of the data on zenodo, default is set to 0.0.0.91}

\item{write_to}{str directory location where to download the data to}
}
\value{
NULL the minted data will be written to the ZENODO_DIR
}
\description{
Download the input data from zenodo
}
