% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ideal.R
\name{generate_idealized_files}
\alias{generate_idealized_files}
\title{Module that prepares all of the Hector inputs for idealized runs}
\usage{
generate_idealized_files(start_yr = 1760)
}
\arguments{
\item{start_yr}{int year indicating the start year of the idealized experiment}
}
\value{
nothing writes out ini file
}
\description{
Module that prepares all of the Hector inputs for idealized runs
}
