% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fxns.R
\name{generate_input_tables}
\alias{generate_input_tables}
\title{Generate emission and concentration constraint input tables for Hector.}
\usage{
generate_input_tables(scenarios, output_dir, years)
}
\arguments{
\item{scenarios}{a character vector of scenario names.}

\item{output_dir}{a character of the full path to the directory location for
the user to write the generated csv files to.}

\item{years}{a vector of the expected output years.}
}
\value{
the location of the generated csv files.
}
\description{
Generate emission and concentration constraint input tables for Hector.
}
