% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rcmip.R
\name{generate_rcmip_rcps}
\alias{generate_rcmip_rcps}
\title{Generate the rcmip rcp inputs, the ini files and csv tables}
\usage{
generate_rcmip_rcps()
}
\value{
nothing writes out the csv and ini files
}
\description{
Generate the rcmip rcp inputs, the ini files and csv tables
}
