% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rcmip.R
\name{generate_rcmip_ssps}
\alias{generate_rcmip_ssps}
\title{Generate the rcmip ssp inputs, the ini files and csv tables}
\usage{
generate_rcmip_ssps()
}
\value{
nothing writes out the csv and ini files
}
\description{
Generate the rcmip ssp inputs, the ini files and csv tables
}
