% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmip_fxns.R
\name{generate_rcmip_submission_files}
\alias{generate_rcmip_submission_files}
\title{Generate the Hector input tables and ini files that were used in RCMIP phases 1 & 2}
\usage{
generate_rcmip_submission_files(
  scenarios_to_process = NULL,
  depends_on = c("rcmip_data.csv")
)
}
\arguments{
\item{scenarios_to_process}{string vector of the scenario names to process}

\item{depends_on}{string vector of the required intermeidate data tables}
}
\value{
nothing writes out the csv and ini files
}
\description{
See https://doi.org/10.5194/gmd-13-5175-2020 for more details
}
\details{
TODO needs better way to determine dependencies
}
