% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v25_fxns.R
\name{generate_v25_files}
\alias{generate_v25_files}
\title{Update the ini files and csv tables that were used with the v2.5 Hector release}
\usage{
generate_v25_files(
  scenarios_to_process = NULL,
  depends_on = c("hectorv2_data.csv")
)
}
\arguments{
\item{scenarios_to_process}{string vector of the scenario names to process}

\item{depends_on}{string vector of the required intermediate data tables}
}
\value{
nothing writes out the csv and ini files
}
\description{
TODO needs better way to determine dependencies
}
