% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hectordata-package.R
\docType{package}
\name{hectordata-package}
\alias{hectordata}
\alias{hectordata-package}
\title{hectordata: Prepare and Format the Input Data for Hector, the Simple Climate Model}
\description{
This pacakge prepares, documents, and formats the inputs for Hector.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JGCRI/hectordata}
  \item Report bugs at \url{https://github.com/JGCRI/hectordata/issues}
}

}
\author{
\strong{Maintainer}: Kalyn Dorheim \email{kalyn.dorheim@pnnl.gov} (\href{https://orcid.org/0000-0001-8093-8397}{ORCID})

Other contributors:
\itemize{
  \item Matt Nicholson \email{matthew.nicholson@pnnl.gov} [contributor]
  \item Ben Bond-Lamberty \email{BondLamberty@pnnl.gov} [contributor]
  \item Chris Vernon \email{chris.vernon@pnnl.gov} [contributor]
  \item Alexey Shiklomanov \email{alexey.shiklomanov@gmail.com} [contributor]
}

}
\keyword{internal}
