% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ideal.R
\name{idealized_ini_table_maker}
\alias{idealized_ini_table_maker}
\title{A helper function that formats and write the idealized hector ini files out}
\usage{
idealized_ini_table_maker(dat, csv_path)
}
\arguments{
\item{dat}{data fame containing some information about corresponding csv file the scenario and  variables}

\item{csv_path}{string vector of the required intermediate data tables}
}
\value{
nothing writes out ini file
}
\description{
This works for a singel scenario at a time. If multiple scneairos are read in it will
throw and error message!

TODO should not be writing a custom table maker function for each set of ini files but here we are
}
