% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ini_template}
\alias{ini_template}
\title{template_ini}
\format{
character
\describe{ A character vector where each element vector contains a different
line of hector ini file.
}
}
\usage{
ini_template
}
\description{
Character vector of the ini lines
}
\keyword{datasets}
