% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns_helper.R
\name{load_data}
\alias{load_data}
\title{Read in hectordata data system inputs}
\usage{
load_data(files)
}
\arguments{
\item{files}{vector of the csv files names needed for the module}
}
\value{
list of the data frame that will be used in the module
}
\description{
Read in hectordata data system inputs
}
