% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ideal.R
\name{make_hector_1pctCO2_df}
\alias{make_hector_1pctCO2_df}
\title{Make the data frame of the Hector inputs for a concentration driven abrupt 1pct CO2 run}
\usage{
make_hector_1pctCO2_df(first_yr)
}
\arguments{
\item{first_yr}{int year indicating the start year of the idealized experiment}
}
\value{
dataframe of the inputs, using constant constant emissions and concentration constraints when appropriate.
}
\description{
This function produces the table of Hector inputs for the 1pct CO2 run
inputs. Any ghg that has a pre-industrial concentration must be held constant at the value
in order to have zero radiative forcing effect. The CO2 concentration increases by 1\% per year.
}
