% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ideal.R
\name{make_hector_abruptx2CO2_df}
\alias{make_hector_abruptx2CO2_df}
\title{Make the data frame of the Hector inputs for a concentration driven abrupt 2 x step}
\usage{
make_hector_abruptx2CO2_df(first_yr)
}
\arguments{
\item{first_yr}{int year indicating the start year of the idealized experiment}
}
\value{
dataframe of the inputs, using constant constant emissions and concentration constraints when appropriate.
}
\description{
This function produces the table of Hector inputs for the abrupt 2 x step test run
inputs. Any ghg that has a pre-industrial concentration must be held constant at the value
in order to have zero radiative forcing effect. The CO2 concentration constraint has
is held constant until the \link{first_yr} when the concentration will double.
}
