% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ideal.R
\name{make_hector_picontrol_df}
\alias{make_hector_picontrol_df}
\title{Make the data frame of the Hector pre industrial control inputs}
\usage{
make_hector_picontrol_df()
}
\value{
dataframe of the pre industrial control inputs, uses constant concentration constraints when appropriate.
}
\description{
This function produces the table of Hector inputs for the preindustrial control run
inputs. Any ghg that has a preindustrial concentration must be held constant at the value
in order to have zero radiative forcing effect.
}
