% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-ini.R
\name{make_new_ini}
\alias{make_new_ini}
\title{Generate a new ini file for a given emissions file}
\usage{
make_new_ini(files, iniprefix_ = NULL)
}
\arguments{
\item{files}{str vector of the new csv hector input files}

\item{iniprefix_}{optional string prefix to add the ini file name}
}
\value{
str of the ini files
}
\description{
Generate a new ini file for a given emissions file
}
