% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rcmip_idealized.R
\name{module_rcmip_idealized}
\alias{module_rcmip_idealized}
\title{Generate the idealized ini & csv files from the rcmip source}
\usage{
module_rcmip_idealized()
}
\value{
str of the ini files
}
\description{
Creates ini and csv tables for the following scenarios based on the RCMIP protocols
1pctCO2, 1pctCO2-4xext, abrupt-0p5xCO2, abrupt-2xCO2, piControl, abrupt-4xCO2, 1pctCO2-cdr
}
