% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rcmip_rcp.R
\name{module_rcmip_rcp}
\alias{module_rcmip_rcp}
\title{Generate the rccp ini & csv files from the rcmip source}
\usage{
module_rcmip_rcp()
}
\value{
str of the ini files
}
\description{
Creates ini and csv tables for the following scenarios based on the RCMIP protocols
rcp26, rcp45, rcp60, rcp85
}
