% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rcmip_ssp.R
\name{module_rcmip_ssp}
\alias{module_rcmip_ssp}
\title{Generate the ssp ini & csv files from the rcmip source}
\usage{
module_rcmip_ssp(...)
}
\value{
str of the ini files
}
\description{
Generate the ssp ini & csv files from the rcmip source
}
