% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ini.R
\name{parse_ini_fname}
\alias{parse_ini_fname}
\title{Parse the filename of a Hector input .ini file for a given scenario}
\usage{
parse_ini_fname(scenario)
}
\arguments{
\item{scenario}{Character vector; Name of the scenario.}
}
\value{
Character vector; Name of the emissions file corresponding to the scenario
}
\description{
The filename will be of the format: "hector_<scenario>.ini"
}
\keyword{internal}
