% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_carbon_cycle_emissions}
\alias{process_carbon_cycle_emissions}
\title{Formate the carbon cycle emissions, they must be postive values}
\usage{
process_carbon_cycle_emissions(dat)
}
\arguments{
\item{dat}{a data table emissions & concentration data}
}
\value{
emissions and concentrations data frame with the correctly formatted carbon cycle emissions aka no negative emissions
}
\description{
Formate the carbon cycle emissions, they must be postive values
}
