% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rcmip.R, R/rcmip_fxns.R
\name{process_rcmip_data}
\alias{process_rcmip_data}
\title{Process the inputs materials from RCMIP into the appropriate units for Hector}
\usage{
process_rcmip_data(scenarios_to_process = NULL)

process_rcmip_data(scenarios_to_process = NULL)
}
\arguments{
\item{scenarios_to_process}{string vector of the scenario names to process}
}
\value{
data frame of Hector inputs for multiple scenarios

data frame of Hector inputs for multiple scenarios
}
\description{
Process the inputs materials from RCMIP into the appropriate units for Hector

Process the inputs materials from RCMIP into the appropriate units for Hector
}
