% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v25_fxns.R
\name{process_v25_data}
\alias{process_v25_data}
\title{Process the inputs materials from Hector v2.5}
\usage{
process_v25_data(scenarios_to_process = NULL)
}
\arguments{
\item{scenarios_to_process}{string vector of the scenario names to process}
}
\value{
data frame of Hector inputs for multiple scenarios
}
\description{
See release Hector v2.5 commit 62381e7
}
