% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ini.R
\name{replace_ini_vars}
\alias{replace_ini_vars}
\title{Replace the placeholder variables in the template Hector .ini file with the
propper scenario variables}
\usage{
replace_ini_vars(scenario, emissions_file)
}
\arguments{
\item{scenario}{Character vector; Name of the scenario for which an input file is being created.}

\item{emission_file}{Character vector; Name of the emission file for the given scenario.}
}
\value{
Character vector; Lines of the scenario .ini
}
\description{
Helper function for create_scenario_ini
}
\keyword{internal}
