% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fxns.R
\name{save_output}
\alias{save_output}
\title{Format the Hector input into the expected Hector input csv file}
\usage{
save_output(x, filename)
}
\arguments{
\item{x}{a data table containing the Hector input information}

\item{filename}{character path for where to save the output.}
}
\value{
csv files formatted to look like hector input tables.
}
\description{
Format the Hector input into the expected Hector input csv file
}
