% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ini.R
\name{write_file}
\alias{write_file}
\title{Write a character vector to file}
\usage{
write_file(file_lines, out_path)
}
\arguments{
\item{file_lines}{Character vector; Lines of the file to write}

\item{out_path}{Character vector; Path of the file to write}
}
\description{
Write a character vector to file
}
\keyword{internal}
