% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxns-tables.R
\name{write_hector_csv}
\alias{write_hector_csv}
\title{Save the hector csv files into the proper hector format}
\usage{
write_hector_csv(
  x,
  required,
  write_to,
  info_source,
  end_tag = "_emiss-constraints_rf"
)
}
\arguments{
\item{x}{data table containing Hector input values}

\item{required}{str vector of the required variables that must be included in the table}

\item{write_to}{str directory to write the hector csv output to}

\item{info_source}{string name indicating the source part of the module name}

\item{end_tag}{string used at the end of of the table file, in most cases will be default name "_emiss-constraints_rf"}
}
\value{
str file name
}
\description{
Save the hector csv files into the proper hector format
}
