% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ini.R
\name{parse_emission_fname}
\alias{parse_emission_fname}
\title{Parse the name of the emissions file for a given scenario.}
\usage{
parse_emission_fname(scenario)
}
\arguments{
\item{scenario}{Character vector; Name of the scenario.}
}
\value{
Character vector; Name of the emissions file corresponding to the scenario
}
\description{
The filename will be of the format: "<scenario>_emissions.csv".
Note: the letters in the scenario name will be converted to upper-case to follow
existing Hector input file naming conventions.
}
\examples{
parse_emission_fname("rcp45")

}
\keyword{internal}
