% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_ini}
\alias{template_ini}
\title{template_ini}
\format{
character
\describe{ A character vector where each element vector contains a different
line of hector ini file.
}
}
\usage{
template_ini
}
\description{
List of data.tables containing information that is used to convert input data provided from
various sources such as IIASA, GCAM, RCMIP and so on into the appropriate values for Hector.
}
\keyword{datasets}
